+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## main
 
+- Add `config.fastlane_broadcasts_enabled` to opt-in for optimized broadcasts (no double JSON encoding).
+
 ## 0.1.2
 
 - Added a hack to prevent third-party extensions from changing the methods visibility.